ADD: Tensorflow Serving Benchmarking (#307)
