+++ tests/unit/test_launch_benchmark.py
@@ -37,6 +37,9 @@ test_precision = "fp32"
 test_docker_image = "foo"
 test_batch_size = "100"
 test_num_cores = "1"
+# need a valid file for tests to work, see conftest.py for where this is managed
+test_input_graph = "test.pb"
+test_tfserving_framework = "tensorflow_serving"
 
 
 @pytest.fixture
@@ -66,8 +69,35 @@ def mock_system_platform(patch):
     return patch("common.base_benchmark_util.platform_util.system_platform")
 
 
+@pytest.fixture
+def mock_path_exists(patch):
+    return patch("os.path.exists", MagicMock(return_value=True))
+
+
+@pytest.fixture
+def mock_isfile(patch):
+    return patch("os.path.isfile", MagicMock(return_value=True))
+
+
+@pytest.fixture
+def mock_isdir(patch):
+    return patch("os.path.isdir", MagicMock(return_value=True))
+
+
+@pytest.fixture
+def mock_islink(patch):
+    return patch("os.path.islink", MagicMock(return_value=False))
+
+
+@pytest.fixture
+def mock_stat(patch):
+    stat = MagicMock()
+    stat.return_value.st_nlink = 0
+    return patch("os.stat", stat)
+
+
 @pytest.fixture(autouse=True)
-def launch_benchmark(mock_platform_util, request):
+def launch_benchmark(mock_platform_util, request, mock_isdir, mock_isfile, mock_islink, mock_stat, mock_path_exists):
     """sets up launch_benchmark obj for every test case and handles catching errors if we wanna test that
        To catch errors called when running launch_benchmark, use something like:
            ['catch_error', SystemExit, [{args}], {error_message}] in parametrize
@@ -113,8 +143,10 @@ def launch_benchmark(mock_platform_util, request):
                 req_args = request.param[2]
             error_message = request.param[3] if len(request.param) == 4 else ''
         else:
+            # add extra arguments to the default ones when calling LaunchBenchmark
             req_args = request.param + example_req_args
     else:
+        # only use default arguments when calling LaunchBenchmark
         req_args = example_req_args
 
     with mock_patch.object(sys, "argv", ['run_tf_benchmark.py'] + req_args):
@@ -168,12 +200,13 @@ def test_launch_benchmark_parse_unknown_args(launch_benchmark):
                                                                            "--output-results"],
                                                   "--output-results can only be used when running "
                                                   "inference with a dataset"],
-                                              ['catch_error', SystemExit, ["--model-name", test_model_name,
-                                                                           "--framework", test_framework,
-                                                                           "--mode", test_mode,
-                                                                           "--precision", test_precision,
-                                                                           "--volume", "~:test"],
-                                                  "Volume mounts can only be used when running in a docker container"],
+                                              ['catch_error_override_all_params', SystemExit,
+                                               ["--model-name", test_model_name,
+                                                "--framework", test_framework,
+                                                "--mode", test_mode,
+                                                "--precision", test_precision,
+                                                "--volume", "~:test"],
+                                                  "Volume mounts can only be used when running in a docker container"]
                                               ], indirect=True)
 def test_launch_benchmark_parse_bad_args(launch_benchmark):
     """
@@ -214,6 +247,31 @@ def test_bare_metal(launch_benchmark, mock_popen):
     launch_benchmark.run_bare_metal("/foo", "/bar", test_env_vars)
     assert mock_popen.called
     args, kwargs = mock_popen.call_args
+
+    # make sure that the start script is run
+    assert "bash" == args[0][0]
+    assert "start.sh" in args[0][1]
+
+    # ensure env vars are set
+    assert os.environ["TEST_ENV_VAR_1"] == test_env_vars["TEST_ENV_VAR_1"]
+    assert os.environ["TEST_ENV_VAR_2"] == test_env_vars["TEST_ENV_VAR_2"]
+
+
+@pytest.mark.parametrize('launch_benchmark', [["--in-graph", test_input_graph]], indirect=True)
+def test_launch_benchmark_tensorflow_serving_framework(launch_benchmark, mock_popen):
+    """
+    Tests that the launch script works for tensorflow serving framework
+    """
+    test_env_vars = {"TEST_ENV_VAR_1": "a", "TEST_ENV_VAR_2": "b"}
+    # Override framework and docker image.
+    launch_benchmark.args.framework = test_tfserving_framework
+    launch_benchmark.args.docker_image = None
+    launch_benchmark.run_bare_metal("/foo", "/bar", test_env_vars)
+    assert mock_popen.called
+    args, kwargs = mock_popen.call_args
+
+    assert launch_benchmark.args.input_graph == test_input_graph
+    assert launch_benchmark.args.framework == test_tfserving_framework
 
     # make sure that the start script is run
     assert "bash" == args[0][0]
