+++ tox.ini
@@ -37,6 +37,7 @@ omit =
   .tox/*
   .pytest_cache/*
   __pycache__/*
+  benchmarks/image_recognition/tensorflow_serving/*
   benchmarks/image_segmentation/tensorflow/unet/inference/fp32/unet_infer.py
   benchmarks/object_detection/tensorflow/ssd-mobilenet/inference/fp32/infer_detections.py
   benchmarks/recommendation/tensorflow/wide_deep/inference/fp32/data_download.py
